move segments out of mill.runner.api
