+++ core/define/src/mill/define/Segment.scala
@@ -1,4 +1,6 @@
-package mill.runner.api
+package mill.define
+
+
 sealed trait Segment {
   def pathSegments: Seq[String] = this match {
     case Segment.Label(s) => Seq(s)
