+++ core/define/src/mill/define/Segments.scala
@@ -1,4 +1,6 @@
-package mill.runner.api
+package mill.define
+
+
 import scala.math.Ordering.Implicits.seqOrdering
 
 /**
