+++ core/define/src/mill/define/package.scala
@@ -3,8 +3,4 @@ package mill
 package object define {
   val SelectMode = mill.runner.api.SelectMode
   type SelectMode = mill.runner.api.SelectMode
-  val Segment = mill.runner.api.Segment
-  type Segment = mill.runner.api.Segment
-  val Segments = mill.runner.api.Segments
-  type Segments = mill.runner.api.Segments
 }
