+++ core/exec/src/mill/exec/Execution.scala
@@ -27,7 +27,7 @@ private[mill] case class Execution(
     rootModule: BaseModuleApi,
     classLoaderSigHash: Int,
     classLoaderIdentityHash: Int,
-    workerCache: mutable.Map[Segments, (Int, Val)],
+    workerCache: mutable.Map[String, (Int, Val)],
     env: Map[String, String],
     failFast: Boolean,
     threadCount: Option[Int],
@@ -45,7 +45,7 @@ private[mill] case class Execution(
            rootModule: BaseModuleApi,
            classLoaderSigHash: Int,
            classLoaderIdentityHash: Int,
-           workerCache: mutable.Map[Segments, (Int, Val)],
+           workerCache: mutable.Map[String, (Int, Val)],
            env: Map[String, String],
            failFast: Boolean,
            threadCount: Option[Int],
